producer/requirements.txt
kafka-python
firebase_admin